wondertask/requirements.txt
asgiref==3.3.4
Django==3.2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
Pillow==8.2.0
sqlparse==0.4.1
django-taggit==1.4.0
django-taggit-serializer==0.1.7
django-mptt==0.12.0
psycopg2-binary==2.8.6
django-filter==2.4.0
celery==5.0.5
redis==3.5.3
pytest-django==4.4.0
django-celery-beat==2.2.0
pytest-celery==0.0.0
